Snapshots without a timestamp now have a .0 extension.
If a synchronizing process is interrupted and resumed some time later,
it is not relevant for it to have a timestamp before it is completed.
Until the timestamp property is assigned a datetime value, it is None
and the path extension is .0 in order to appear first in sorted
directory listings.