Made the definition of the _logger attribute DRY.
Most classes in this program define a _logger attribute.
This is now done in a Logging class of the _logging module. Classes must
simply subclass it and start using methods of self._logger.

This makes future changes to the loggers naming scheme a simple matter
of modifying the Logging class.