Bugfix: Corrected Cycle.purge algorithm.
Formerly, the algorithm took the maxnumber argument and incremented it
by 1 for each incomplete snapshot in the list, then deleted the excess
snapshots. This would leave "dangling" snapshots at the tail of the
list.

Now, as soon as maxnumber complete snapshots are counted, the remaining
snapshots are deleted, as well as any trailing incomplete snapshots.