Made Cycle interruptible by PipeLogger thread through an event object.
While Cycle instances wait for the rsync subprocess to finish, it also
monitors the kill_switch_event that might be set by a PipeLogger
instance, and also handles KeyboardInterrupt. In both cases, it kills
the rsync subprocess.

The interrupt_event instance attribute of the PipeLogger class is
removed. Logging must continue until the end of the stream. There is no
need to talk to the thread if the subprocess is killed and it stops
outputting to stderr and stdout.